Moved range_concept to its own header
